requirements.txt
emoji
nltk
transformers
pytorch_lightning
tensorboard
tqdm